requirements.txt
faiss-cpu
sentence-transformers
google-generativeai
PyPDF2
numpy
pandas
tqdm
